Update utils.js
When I was trying to use the ORM to create tables in postgreSQL with some column as 'varying character' datatype, I found that all string data types are converted into TEXT, even if I have given the size as specified in the documentation, the schema was not creating properly. So I made some changes in the case handling portion in utils.js file and it started working. If the size handling changes got added, it will be a great help for everyone. I even checked with the mysql library , there this particular issue is not present. Thanks.